created delete button and pulled main fragment into is own file
