+++ app/src/main/res/values/strings.xml
@@ -47,4 +47,7 @@
     <string name="pref_amount_units_litres" translatable="false">Liters</string>
     <string name="pref_amount_units_us_gallon" translatable="false">US gallons</string>
     <string name="pref_amount_units_imperial_gallon" translatable="false">Imperial gallons</string>
+
+<!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
